Moving function to console app
