+++ tests/test_cache_lock_file_perf.py
@@ -1,14 +1,11 @@
-import logging
 import multiprocessing
 import os
 import shutil
 import tempfile
-import time
 
-from portalocker import exceptions
 import pytest
 
-from msal_extensions import FilePersistence, CrossPlatLock
+from tests.cache_file_generator import _acquire_lock_and_write_to_cache
 
 
 @pytest.fixture
@@ -35,22 +32,6 @@ def _validate_result_in_cache(cache_location):
                 assert tag == '<', "Opening bracket not found"
                 prev_process_id = process_id
     return count
-
-
-def _acquire_lock_and_write_to_cache(cache_location, sleep_interval):
-    cache_accessor = FilePersistence(cache_location)
-    lock_file_path = cache_accessor.get_location() + ".lockfile"
-    try:
-        with CrossPlatLock(lock_file_path):
-            data = cache_accessor.load()
-            if data is None:
-                data = ""
-            data += "< " + str(os.getpid()) + "\n"
-            time.sleep(sleep_interval)
-            data += "> " + str(os.getpid()) + "\n"
-            cache_accessor.save(data)
-    except exceptions.LockException as e:
-        logging.warning("Unable to acquire lock %s", e)
 
 
 def _run_multiple_processes(no_of_processes, cache_location, sleep_interval):
